Bug fixes
Bug fixes for the Register redirect, and the add contact fields and button was missing and has been fixed. A couple of variables were commented out for the purpose that they were not being used. For ease of finding any other problems caused by my editing of the code. I have commented my name by the snippet of code changed and the date of the change.